pyproject.toml
[project]
name = "crudgen"
version = "0.1.0"
description = "Config-driven CRUD generator for FastAPI (no runtime magic)"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
crudgen = "crudgen.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
